project structure refactoring. src
