+++ mdb/lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:mdb/mysrc/app.dart';
+import 'package:mdb/src/app.dart';
 
 void main() {
 //  debugPaintSizeEnabled=true; // visualize layout fo debug
