+++ mdb/lib/src/data/repository/movie_repository.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
-import 'package:mdb/mysrc/data/api/api.dart';
-import 'package:mdb/mysrc/data/model/remote/discover_response.dart';
-import 'package:mdb/mysrc/data/model/remote/genres_response.dart';
-import 'package:mdb/mysrc/data/model/remote/popular_response.dart';
+import 'package:mdb/src/data/api/api.dart';
+import 'package:mdb/src/data/model/remote/discover_response.dart';
+import 'package:mdb/src/data/model/remote/genres_response.dart';
+import 'package:mdb/src/data/model/remote/popular_response.dart';
 
 class MovieRepository {
   final api = Api();
