+++ mdb/lib/src/ui/home_screen.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:mdb/mysrc/data/model/local/genre.dart';
-import 'package:mdb/mysrc/data/model/local/movie.dart';
-import 'package:mdb/mysrc/data/model/remote/discover_response.dart';
-import 'package:mdb/mysrc/data/model/remote/genres_response.dart';
-import 'package:mdb/mysrc/data/model/remote/popular_response.dart';
-import 'package:mdb/mysrc/ui/movie_page_viver_item.dart';
-import 'package:mdb/mysrc/utils/constants.dart';
-import 'package:mdb/mysrc/utils/wigdet/page_transformer.dart';
+import 'package:mdb/src/data/model/local/genre.dart';
+import 'package:mdb/src/data/model/local/movie.dart';
+import 'package:mdb/src/data/model/remote/discover_response.dart';
+import 'package:mdb/src/data/model/remote/genres_response.dart';
+import 'package:mdb/src/data/model/remote/popular_response.dart';
+import 'package:mdb/src/ui/movie_page_viver_item.dart';
+import 'package:mdb/src/utils/constants.dart';
+import 'package:mdb/src/utils/wigdet/page_transformer.dart';
 
 class HomeScreen extends StatefulWidget {
   @override
